pyproject.toml
[project]
name = "baisc-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.7.1.1",
    "openai-agents>=0.2.9",
]
